requirements.txt
Flask==3.1.2
groq==0.31.1
python-dotenv==1.1.1
psycopg2-binary==2.9.10
requests==2.32.5
gunicorn==23.0.0

